Move command completer and prompt functions to cli_prompts.py and move /settings command functionality to cli_settings.py
