+++ openhands/core/cli_settings.py
@@ -1,0 +1,236 @@
+from prompt_toolkit import PromptSession, print_formatted_text
+from prompt_toolkit.completion import FuzzyWordCompleter
+from prompt_toolkit.formatted_text import HTML
+from pydantic import SecretStr
+
+import openhands.agenthub  # noqa F401 (we import this to get the agents registered)
+from openhands.controller.agent import Agent
+from openhands.core.cli_prompts import UserCancelledError, cli_confirm, kb_cancel
+from openhands.core.cli_utils import (
+    VERIFIED_ANTHROPIC_MODELS,
+    VERIFIED_OPENAI_MODELS,
+    VERIFIED_PROVIDERS,
+    organize_models_and_providers,
+)
+from openhands.core.config import AppConfig
+from openhands.core.config.condenser_config import NoOpCondenserConfig
+from openhands.core.config.utils import OH_DEFAULT_AGENT
+from openhands.memory.condenser.impl.llm_summarizing_condenser import (
+    LLMSummarizingCondenserConfig,
+)
+from openhands.storage.data_models.settings import Settings
+from openhands.storage.settings.file_settings_store import FileSettingsStore
+from openhands.utils.llm import get_supported_llm_models
+
+
+async def modify_llm_settings_basic(
+    config: AppConfig, settings_store: FileSettingsStore
+) -> bool:
+    model_list = get_supported_llm_models(config)
+    organized_models = organize_models_and_providers(model_list)
+
+    provider_list = list(organized_models.keys())
+    verified_providers = [p for p in VERIFIED_PROVIDERS if p in provider_list]
+    provider_list = [p for p in provider_list if p not in verified_providers]
+    provider_list = verified_providers + provider_list
+
+    provider_completer = FuzzyWordCompleter(provider_list)
+    session = PromptSession(key_bindings=kb_cancel())
+
+    provider = None
+    model = None
+    api_key = None
+
+    try:
+        provider = await session.prompt_async(
+            '(Step 1/3) Select LLM Provider (use Tab for completion): ',
+            completer=provider_completer,
+        )
+
+        if provider not in organized_models:
+            print_formatted_text(
+                HTML(f'\n<grey>Invalid provider selected: {provider}</grey>\n')
+            )
+            return False
+
+        model_list = organized_models[provider]['models']
+        if provider == 'openai':
+            model_list = [m for m in model_list if m not in VERIFIED_OPENAI_MODELS]
+            model_list = VERIFIED_OPENAI_MODELS + model_list
+        if provider == 'anthropic':
+            model_list = [m for m in model_list if m not in VERIFIED_ANTHROPIC_MODELS]
+            model_list = VERIFIED_ANTHROPIC_MODELS + model_list
+
+        model_completer = FuzzyWordCompleter(model_list)
+        model = await session.prompt_async(
+            '(Step 2/3) Select LLM Model (use Tab for completion): ',
+            completer=model_completer,
+        )
+
+        if model not in organized_models[provider]['models']:
+            print_formatted_text(
+                HTML(
+                    f'\n<grey>Invalid model selected: {model} for provider {provider}</grey>\n'
+                )
+            )
+            return False
+
+        session.completer = None  # Reset completer for password prompt
+        api_key = await session.prompt_async('(Step 3/3) Enter API Key: ')
+
+    except (
+        UserCancelledError,
+        KeyboardInterrupt,
+        EOFError,
+    ):
+        return False  # Return False on exception
+
+    # TODO: check for empty string inputs?
+    # Handle case where a prompt might return None unexpectedly
+    if provider is None or model is None or api_key is None:
+        return False
+
+    save_settings = (
+        cli_confirm(
+            '\nSave new settings? Current session will be terminated!',
+            ['Yes, proceed', 'No, dismiss'],
+        )
+        == 0
+    )
+
+    if not save_settings:
+        return False
+
+    llm_config = config.get_llm_config()
+    llm_config.model = provider + organized_models[provider]['separator'] + model
+    llm_config.api_key = SecretStr(api_key)
+    llm_config.base_url = None
+    config.set_llm_config(llm_config)
+
+    config.default_agent = OH_DEFAULT_AGENT
+    config.security.confirmation_mode = False
+    config.enable_default_condenser = True
+
+    agent_config = config.get_agent_config(config.default_agent)
+    agent_config.condenser = LLMSummarizingCondenserConfig(
+        llm_config=llm_config,
+        type='llm',
+    )
+    config.set_agent_config(agent_config, config.default_agent)
+
+    settings = await settings_store.load()
+    if not settings:
+        settings = Settings()
+
+    settings.llm_model = provider + organized_models[provider]['separator'] + model
+    settings.llm_api_key = SecretStr(api_key)
+    settings.llm_base_url = None
+    settings.agent = OH_DEFAULT_AGENT
+    settings.confirmation_mode = False
+    settings.enable_default_condenser = True
+
+    await settings_store.store(settings)
+
+    return True
+
+
+async def modify_llm_settings_advanced(
+    config: AppConfig, settings_store: FileSettingsStore
+) -> bool:
+    session = PromptSession(key_bindings=kb_cancel())
+
+    custom_model = None
+    base_url = None
+    api_key = None
+    agent = None
+
+    try:
+        custom_model = await session.prompt_async('(Step 1/6) Custom Model: ')
+        base_url = await session.prompt_async('(Step 2/6) Base URL: ')
+        api_key = await session.prompt_async('(Step 3/6) API Key: ')
+
+        agent_list = Agent.list_agents()
+        agent_completer = FuzzyWordCompleter(agent_list)
+        agent = await session.prompt_async(
+            '(Step 4/6) Agent (use Tab for completion): ', completer=agent_completer
+        )
+
+        if agent not in agent_list:
+            print_formatted_text(
+                HTML(f'\n<grey>Invalid agent selected: {agent}</grey>\n')
+            )
+            return False
+
+        enable_confirmation_mode = (
+            cli_confirm(
+                question='(Step 5/6) Confirmation Mode:',
+                choices=['Enable', 'Disable'],
+            )
+            == 0
+        )
+
+        enable_memory_condensation = (
+            cli_confirm(
+                question='(Step 6/6) Memory Condensation:',
+                choices=['Enable', 'Disable'],
+            )
+            == 0
+        )
+
+    except (
+        UserCancelledError,
+        KeyboardInterrupt,
+        EOFError,
+    ):
+        return False  # Return False on exception
+
+    # TODO: check for empty string inputs?
+    # Handle case where a prompt might return None unexpectedly
+    if custom_model is None or base_url is None or api_key is None or agent is None:
+        return False
+
+    save_settings = (
+        cli_confirm(
+            '\nSave new settings? Current session will be terminated!',
+            ['Yes, proceed', 'No, dismiss'],
+        )
+        == 0
+    )
+
+    if not save_settings:
+        return False
+
+    llm_config = config.get_llm_config()
+    llm_config.model = custom_model
+    llm_config.base_url = base_url
+    llm_config.api_key = SecretStr(api_key)
+    config.set_llm_config(llm_config)
+
+    config.default_agent = agent
+
+    config.security.confirmation_mode = enable_confirmation_mode
+
+    agent_config = config.get_agent_config(config.default_agent)
+    if enable_memory_condensation:
+        agent_config.condenser = LLMSummarizingCondenserConfig(
+            llm_config=llm_config,
+            type='llm',
+        )
+    else:
+        agent_config.condenser = NoOpCondenserConfig(type='noop')
+    config.set_agent_config(agent_config)
+
+    settings = await settings_store.load()
+    if not settings:
+        settings = Settings()
+
+    settings.llm_model = custom_model
+    settings.llm_api_key = SecretStr(api_key)
+    settings.llm_base_url = base_url
+    settings.agent = agent
+    settings.confirmation_mode = enable_confirmation_mode
+    settings.enable_default_condenser = enable_memory_condensation
+
+    await settings_store.store(settings)
+
+    return True
